Refactor virtual fs into a trait object to allow for multiple implementations
